Revert "Merge branch 'main' of https://github.com/headwirecom/helix-sportsmagazine"
This reverts commit a10b7eba1ba3ff0f7a3790859acc0fdadb9813ac, reversing
changes made to bc3131749b6019bc2aff96bd60ca0b8a18534fcf.
