+++ 404.html
@@ -1,190 +1,67 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <title>404 | GolfDigest.com</title>
-    <script type="text/javascript">
-      window.isErrorPage = true;
-      window.errorCode = "404";
-    </script>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta property="og:title" content="404 | GolfDigest.com" />
-    <script
-      src="/scripts/scripts.js"
-      type="module"
-      crossorigin="use-credentials"
-    ></script>
-    <link rel="stylesheet" href="/styles/styles.css" />
-    <link rel="stylesheet" href="/styles/lazy-styles.css" />
-    <link rel="stylesheet" href="/styles/fonts.css" />
-    <style>
-      main {
-        padding: 0;
-      }
 
-      section {
-        padding: 50px 0;
-        background-color: #f6f6f6;
-      }
+<head>
+  <title>Page not found</title>
+  <script type="text/javascript">
+    window.isErrorPage = true;
+    window.errorCode = '404';
+  </script>
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <meta property="og:title" content="Page not found">
+  <script src="/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
+  <script type="module">
+    import { sampleRUM } from '/scripts/scripts.js';
 
-      .error-container {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        padding-bottom: 50px;
-        flex-wrap: wrap-reverse;
-      }
-
-      .error-message h1 {
-        font-family: var(--font-tungsten);
-        font-weight: 600;
-        font-size: 89px;
-        line-height: 78.32px;
-        letter-spacing: 4.45px;
-        text-transform: uppercase;
-        margin: 0;
-        margin-bottom: 17px;
-      }
-
-      .error-message p {
-        font-family: var(--font-gotham);
-        line-height: 25.4px;
-        text-align: left;
-        color: #8f8f8f;
-        margin: 0;
-        margin-bottom: 26px;
-        opacity: 0.8;
-        width: 450px;
-      }
-
-      .error-message .button {
-        width: 200px;
-        height: 40px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        background-color: #ed3e49;
-        color: #fff;
-        font-family: var(--font-gotham);
-        font-weight: 700;
-        font-size: 12px;
-        line-height: 18.72px;
-        letter-spacing: 0.3em;
-        text-decoration: none;
-        padding: 0;
-      }
-
-      .error-image-and-number {
-        display: flex;
-      }
-
-      .error-image {
-        position: relative;
-        height: auto;
-      }
-
-      .error-image::after {
-        content: "";
-        width: 100%;
-        background: linear-gradient(180deg, hsla(0, 0%, 90%, 0), #f6f6f6);
-        height: 40%;
-        position: absolute;
-        bottom: 0;
-        left: 0;
-        z-index: 1;
-      }
-
-      .error-number {
-        font-family: var(--font-tungsten);
-        font-weight: 500;
-        font-size: 450px;
-        line-height: 540px;
-        color: hsla(0, 0%, 56%, 0.3);
-        opacity: 0.6;
-        margin-left: -90px;
-      }
-
-      header {
-        height: auto;
-      }
-
-      @media (max-width: 768px) {
-        .error-container {
-          flex-direction: column;
-          align-items: center;
-          padding: 0 14px 240px;
-        }
-
-        .error-message {
-          display: flex;
-          justify-content: center;
-          flex-direction: column;
-          align-items: center;
-          order: 2;
-          margin-top: 90px;
-        }
-
-        .error-message p {
-          text-align: center;
-          width: 365px;
-          padding: 0 30px;
-        }
-
-        .error-message h1 {
-          font-weight: 600;
-          font-size: 50px;
-          line-height: 52px;
-          letter-spacing: 2.5px;
-        }
-
-        .error-image-and-number {
-          display: flex;
-          width: 100%;
-          order: 1;
-        }
-
-        .error-image img {
-          width: 200px;
-        }
-
-        .error-number {
-          font-size: 230px;
-          font-weight: 350;
-          font-style: normal;
-          line-height: normal;
-          margin-left: -80px;
+    window.addEventListener('load', () => {
+      if (document.referrer) {
+        const { origin, pathname } = new URL(document.referrer);
+        if (origin === window.location.origin) {
+          const backBtn = document.createElement('a');
+          backBtn.classList.add('button', 'error-button-back');
+          backBtn.href = pathname;
+          backBtn.textContent = 'Go back';
+          backBtn.title = 'Go back';
+          const btnContainer = document.querySelector('.button-container');
+          btnContainer.append(backBtn);
         }
       }
-    </style>
-  </head>
+      sampleRUM('404', { source: document.referrer, target: window.location.href });
+    });
+  </script>
+  <link rel="stylesheet" href="/styles/styles.css">
+  <style>
+    main.error {
+      min-height: calc(100vh - var(--nav-height));
+      display: flex;
+      align-items: center;
+    }
 
-  <body>
-    <header></header>
-    <main class="error">
-      <section>
-        <div class="error-container">
-          <div class="error-message">
-            <h1>TAKE A MULLIGAN</h1>
-            <p>
-              Sorry, we couldn't find the page you were looking for. Check out
-              our homepage for the latest in golf.
-            </p>
-            <a href="/" class="button">TAKE ME THERE</a>
-          </div>
-          <div class="error-image-and-number">
-            <div class="error-image"></div>
-            <div class="error-number">404</div>
-          </div>
-        </div>
-      </section>
-    </main>
-    <footer></footer>
-    <script type="module">
-      import { createOptimizedPicture } from "./scripts/lib-franklin.js";
-      const imageUrl = "/media_10b8f0c21a329ee5ab764abf34d64a94de757ea2a.png";
-      const imageAlt = "Tiger Woods regretting a bad shot";
-      const pictureElement = createOptimizedPicture(imageUrl, imageAlt);
-      const imageContainer = document.querySelector(".error-image");
-      imageContainer.appendChild(pictureElement);
-    </script>
-  </body>
+    main.error .error-number {
+      width: 100%;
+    }
+
+    main.error .error-number text {
+      font-family: var(--fixed-font-family);
+    }
+  </style>
+  <link rel="stylesheet" href="/styles/lazy-styles.css">
+</head>
+
+<body>
+  <header></header>
+  <main class="error">
+    <div class="section">
+      <svg viewBox="1 0 38 18" class="error-number">
+        <text x="0" y="17">404</text>
+      </svg>
+      <h2 class="error-message">Page Not Found</h2>
+      <p class="button-container">
+        <a href="/" class="button secondary error-button-home">Go home</a>
+      </p>
+    </div>
+  </main>
+  <footer></footer>
+</body>
+
 </html>
