+++ blocks/cards/cards.js
@@ -20,24 +20,25 @@ const placeholderHtml = `
 
 export default async function decorate(block) {
   const id = getBlockId(block);
+  const query = window.store.getQuery(block);
 
-  const cardBlocks = [...document.querySelectorAll(".cards.block[data-block-name='cards']:not(.latest)")];
+  const cardBlocks = [...document.querySelectorAll(".cards.block[data-block-name='cards']")];
   const indexInPage = cardBlocks.findIndex((element) => element.isEqualNode(block));
 
   const isLatestCardBlock = block.classList.contains('latest');
 
-  const cardsTitle = block.querySelector('.cards.block h3, .cards.block h3')?.innerText || '';
+  const cardsTitle = block.querySelector('.cards.block h3, .cards.block h3')?.innerText || (isLatestCardBlock ? 'The Latest' : '');
 
   const gdPlusCards = block.classList.contains('gd');
 
-  const reverse = !(indexInPage <= 0 || indexInPage % 2 === 0);
+  const reverse = !(indexInPage === 0 || indexInPage % 2 === 0);
 
   let cardOffset = 4;
   if (block.classList.contains('hero')) {
     cardOffset = 2;
   } else if (block.classList.contains('latest')) {
     cardOffset = 10;
-  } else if (block.classList.contains('columns')) {
+  } else if (block.classList.contains('colums')) {
     cardOffset = 5;
   }
 
@@ -53,8 +54,7 @@ export default async function decorate(block) {
   document.addEventListener(`query:${id}`, (event) => {
     cardData = event.detail.data;
 
-    // TODO Add support for multiple queries
-    const cardList = cardData.slice(currentBlockIndex, currentBlockIndex + cardOffset);
+    const cardList = cardData?.slice(currentBlockIndex, currentBlockIndex + cardOffset);
 
     const mainCard = cardList[0];
 
@@ -65,17 +65,11 @@ export default async function decorate(block) {
       return `
       <a class="main-card" href="${card.href || card.path}">
         <div class="image-bg">
-          ${createOptimizedPicture(card.image, card.imageAlt || card.title, false, [{ width: '700' }]).outerHTML}
+          ${createOptimizedPicture(card.image, card.imageAlt, false, [{ width: '700' }]).outerHTML}
         </div>
         <div class="main-text-wrapper">
           <div class="section">${card.rubric}</div>
-          ${
-  !gdPlusCards
-    ? ''
-    : `<img loading="lazy" src="/icons/${
-      gdPlusCards ? 'gd-plus-light' : 'gd-plus-dark'
-    }.svg" class="gd-plus-icon-img" alt="Golf Digest Plus Icon" />`
-}
+          ${!gdPlusCards ? '' : `<img loading="lazy" src="/icons/${gdPlusCards ? 'gd-plus-light' : 'gd-plus-dark'}.svg" class="gd-plus-icon-img" alt="Golf Digest Plus Icon" />`}
           <div class="headline"><h3>${card.title}</h3></div>
         </div>
       </a>
@@ -87,15 +81,11 @@ export default async function decorate(block) {
         (card) => `
         <a class="small-card" href="${card.href || card.path}">
           <div class="image-wrapper">
-            ${createOptimizedPicture(card.image, card.imageAlt || card.title, false, [{ width: '350' }]).outerHTML}
+            ${createOptimizedPicture(card.image, card.imageAlt, false, [{ width: '350' }]).outerHTML}
           </div>
           <div class="small-text-wrapper">
             <div class="section">${card.rubric}</div>
-            ${
-  !gdPlusCards && !card?.gdPlus
-    ? ''
-    : '<img loading="lazy" src="/icons/gd-plus-dark.svg" class="gd-plus-icon-img" alt="Golf Digest Plus Icon" />'
-}
+            ${!gdPlusCards && !card?.gdPlus ? '' : '<img loading="lazy" src="/icons/gd-plus-dark.svg" class="gd-plus-icon-img" alt="Golf Digest Plus Icon" />'}
             <div class="headline"><h3>${card.title}</h3></div>
             <div class="date-string">${timeSince(convertExcelDate(card.dateValue))}</div>
           </div>
@@ -107,10 +97,7 @@ export default async function decorate(block) {
     const HTML_TEMPLATE = `
     ${cardsTitle ? `<div class="cards-title-wrapper"><h2 class="cards-title">${cardsTitle}</div>` : ''}
     <div class="card-block-wrapper ${reverse ? 'reverse' : ''}">
-      ${
-  block.classList.contains('hero')
-    ? cardList.map((card) => generateMainCard(card)).join('')
-    : `
+      ${block.classList.contains('hero') ? cardList.map((card) => generateMainCard(card)).join('') : `
         ${generateMainCard()}
         ${!gdPlusCards ? '' : '<div class="gd-cards-wrapper">'}
         <div class="secondary-cards">
@@ -118,8 +105,7 @@ export default async function decorate(block) {
           ${generateSecondaryCards(isLatestCardBlock ? cardList : cardList.splice(1))}
         </div>
         ${!gdPlusCards ? '' : '</div>'}
-    `
-}</div>`;
+    `}</div>`;
 
     // Template rendering
     const template = parseFragment(HTML_TEMPLATE);
@@ -134,5 +120,5 @@ export default async function decorate(block) {
   });
 
   // Trigger query
-  window.store.query(block);
+  window.store.query({ id, query, limit: 60 });
 }
