--- blocks/editors-note/editors-note.css
@@ -1,16 +1,0 @@
-
-.editors-note.block .editors-note-content i {
-  font-size: 16px;
-  font-family: var(--font-gotham);
-  font-weight: 500;
-}
-
-@media (min-width: 769px) {
-  .editors-note.block .editors-note-content i {
-    font-size: 18px;
-  }
-}
-
-.editors-note.block .editors-note-content i a {
-  color: #00e;
-}
