+++ blocks/embed/embed.js
@@ -33,58 +33,28 @@ const embedInstagram = (url) => {
 };
 
 const embedYoutube = (url, autoplay) => {
-  let youtubeUrl = new URL(url.href);
-  let params = youtubeUrl.searchParams;
-  if (autoplay) {
-    params.set('muted', '1');
-    params.set('autoplay', '1');
-  }
+  const usp = new URLSearchParams(url.search);
+  const suffix = autoplay ? '&muted=1&autoplay=1' : '';
+  let vid = encodeURIComponent(usp.get('v'));
+  const embed = url.pathname;
   if (url.origin.includes('youtu.be')) {
-    const [, vid] = url.pathname.split('/');
-    youtubeUrl = new URL(`https://www.youtube.com/embed/${vid}`);
-    params = youtubeUrl.searchParams;
-    params.set('rel', '0'); // Set rel=0 to prevent showing related videos
-    params.set('v', vid); // Set the video ID for the embedded video
-    if (autoplay) {
-      params.set('muted', '1');
-      params.set('autoplay', '1');
-    }
+    [, vid] = url.pathname.split('/');
   }
   const embedHTML = `<div style="left: 0; width: 100%; height: 0; position: relative; padding-bottom: 56.25%;">
-      <iframe src="${youtubeUrl.toString()}" style="border: 0; top: 0; left: 0; width: 100%; height: 100%; position: absolute;" 
+      <iframe src="https://www.youtube.com${vid ? `/embed/${vid}?rel=0&v=${vid}${suffix}` : embed}" style="border: 0; top: 0; left: 0; width: 100%; height: 100%; position: absolute;" 
       allow="autoplay; fullscreen; picture-in-picture; encrypted-media; accelerometer; gyroscope; picture-in-picture" allowfullscreen="" scrolling="no" title="Content from Youtube" loading="lazy"></iframe>
     </div>`;
   return embedHTML;
 };
 
 const embedVimeo = (url, autoplay) => {
-  const vimeoURL = new URL(`https://player.vimeo.com/video${url.pathname}`);
-  const params = vimeoURL.searchParams;
-  if (autoplay) {
-    params.set('muted', '1');
-    params.set('autoplay', '1');
-  }
-  vimeoURL.search = params.toString();
+  const [, video] = url.pathname.split('/');
+  const suffix = autoplay ? '?muted=1&autoplay=1' : '';
   const embedHTML = `<div style="left: 0; width: 100%; height: 0; position: relative; padding-bottom: 56.25%;">
-      <iframe src="${vimeoURL.toString()}" 
+      <iframe src="https://player.vimeo.com/video/${video}${suffix}" 
       style="border: 0; top: 0; left: 0; width: 100%; height: 100%; position: absolute;" 
       frameborder="0" allow="autoplay; fullscreen; picture-in-picture" allowfullscreen  
       title="Content from Vimeo" loading="lazy"></iframe>
-    </div>`;
-  return embedHTML;
-};
-
-const embedBrightcove = (url, autoplay) => {
-  const brightcoveUrl = new URL(url.href);
-  const params = brightcoveUrl.searchParams;
-  if (autoplay) {
-    params.set('muted', '1');
-    params.set('autoplay', '1');
-  }
-  const embedHTML = `<div style="left: 0; width: 100%; height: 0; position: relative; padding-bottom: 56.25%;">
-      <iframe src="${brightcoveUrl.toString()}" 
-      style="border: 0; top: 0; left: 0; width: 100%; height: 100%; position: absolute;" 
-      allowfullscreen="" scrolling="no" title="Content from Brightcove" loading="lazy"></iframe>
     </div>`;
   return embedHTML;
 };
@@ -117,10 +87,6 @@ const loadEmbed = (block, link, autoplay) => {
       match: ['instagram'],
       embed: embedInstagram,
     },
-    {
-      match: ['players.brightcove.net'],
-      embed: embedBrightcove,
-    },
   ];
 
   const config = EMBEDS_CONFIG.find((e) => e.match.some((match) => link.includes(match)));
@@ -136,7 +102,6 @@ const loadEmbed = (block, link, autoplay) => {
 };
 
 export default function decorate(block) {
-  const autoplay = block.classList.contains('autoplay') || !!block.closest('.autoplay.block');
   const placeholder = block.querySelector('picture');
   const link = block.querySelector('a').href;
   block.textContent = '';
@@ -147,14 +112,14 @@ export default function decorate(block) {
     wrapper.innerHTML = '<div class="embed-placeholder-play"><button title="Play"></button></div>';
     wrapper.prepend(placeholder);
     wrapper.addEventListener('click', () => {
-      loadEmbed(block, link, autoplay);
+      loadEmbed(block, link, true);
     });
     block.append(wrapper);
   } else {
     const observer = new IntersectionObserver((entries) => {
       if (entries.some((e) => e.isIntersecting)) {
         observer.disconnect();
-        loadEmbed(block, link, autoplay);
+        loadEmbed(block, link);
       }
     });
     observer.observe(block);
