+++ blocks/full-bleed/full-bleed.js
@@ -1,13 +1,5 @@
 import {
-  assignSlot,
-  parseFragment,
-  render,
-  replaceLinksWithEmbed,
-  parseSectionMetadata,
-  addPhotoCredit,
-  ARTICLE_TEMPLATES,
-  premiumArticleBanner,
-  generateArticleBlocker,
+  assignSlot, parseFragment, render, replaceLinksWithEmbed,
 } from '../../scripts/scripts.js';
 import {
   buildBlock,
@@ -16,39 +8,23 @@ import {
   loadBlocks,
 } from '../../scripts/lib-franklin.js';
 
-/**
- * @param {HTMLDivElement} block
- */
-export default async function decorate(block) {
-  const gdPlusArticle = getMetadata('gdplus').length > 0;
+const rubric = getMetadata('rubric');
+const author = getMetadata('author');
+const authorURL = getMetadata('author-url');
+const publicationDate = getMetadata('publication-date');
+const imageCredit = getMetadata('hero-image-credit');
 
-  const rubric = getMetadata('rubric');
-  const author = getMetadata('author');
-  const authorURL = getMetadata('author-url');
-  const publicationDate = getMetadata('publication-date');
-  const headlineMetadata = parseSectionMetadata(block.querySelector('.template-section-metadata'));
-  // TODO remove once importer fixes photo credit
-  const photoCredit = headlineMetadata?.imageCredit ?? headlineMetadata?.photoCredit;
-
-  // HTML template in JS to avoid extra waterfall for LCP blocks
-  const HTML_TEMPLATE = `
-${!gdPlusArticle ? '' : premiumArticleBanner()}
+// HTML template in JS to avoid extra waterfall for LCP blocks
+const HTML_TEMPLATE = `
 <div class="container">
   <div class="lead">
       <div class="headline">
           <p class="rubric">
-              ${!gdPlusArticle ? '' : `
-                <img width="51" height="19" class="gd-plus-icon light" src="/icons/gd-plus-light.svg" alt="GD Plus Icon" />
-                <img width="51" height="19" class="gd-plus-icon dark" src="/icons/gd-plus-dark.svg" alt="GD Plus Icon" />
-              `}
               <span>${rubric}</span>
           </p>
           <div class="title">
               <slot name="heading"></slot>
-              <slot name="description"></slot>
           </div>
-          <!-- this slot must be wrapped in a p tag otherwise other code will break the block -->
-          <p><slot name="editors-note"></slot></p>
           <div class="byline">
               <div class="attribution">
                   <span>By</span>
@@ -62,15 +38,13 @@ ${!gdPlusArticle ? '' : premiumArticleBanner()}
       <div class="image">
           <slot name="image"></slot>
           <div class="credit">
-          <slot name="caption"></slot>
-          ${photoCredit ? `<span>${photoCredit}</span>` : ''}
+              <span>${imageCredit}</span>
           </div>
       </div>
   </div>
   <div class="container-article">
     <div class="content-wrapper">
         <article class="article-content">
-        <slot name="editors-note"></slot>
             <div class="byline">
                 <div class="attribution">
                     <span>By</span>
@@ -94,6 +68,10 @@ ${!gdPlusArticle ? '' : premiumArticleBanner()}
 </div>
 `;
 
+/**
+ * @param {HTMLDivElement} block
+ */
+export default async function decorate(block) {
   // Template rendering
   const template = parseFragment(HTML_TEMPLATE);
 
@@ -102,46 +80,27 @@ ${!gdPlusArticle ? '' : premiumArticleBanner()}
   // Identify slots
   assignSlot(block, 'heading', 'h1');
   assignSlot(block, 'image', 'picture');
-  const h1Element = block.querySelector('h1');
-  const nextSibling = h1Element?.nextElementSibling;
-  if (nextSibling && nextSibling.tagName === 'P' && !nextSibling.querySelector('picture')) {
-    assignSlot(block, 'description', 'h1 + p:not(:has(picture))');
-  }
 
   // Pre-processing
   const share = buildBlock('social-share', { elems: [] });
   share.setAttribute('slot', 'share');
   block.append(share);
 
-  const editorsNote = block.querySelector('.editors-note');
-  if (editorsNote) {
-    decorateBlock(editorsNote);
-    assignSlot(block, 'editors-note', '.editors-note');
-  }
-
   block.querySelectorAll('p').forEach((p) => {
+    // Convert **text** to bold
+    p.innerHTML = p.innerHTML.replace(/\*\*(.*?)\*\*/g, '<strong>$1</strong>');
+    // Center seperator
     if (p.textContent.includes('• • •')) {
       p.classList.add('center-seperator');
     }
   });
 
-  const picture = block.querySelector('picture');
-
-  const caption = picture.parentElement.nextElementSibling;
-  if (caption && caption.tagName === 'P') {
-    caption.setAttribute('slot', 'caption');
-  }
-
   // Render template
-  // TODO remove ARTICLE_TEMPLATES.FullBleed once importer fixes "**" occurrences
-  render(template, block, ARTICLE_TEMPLATES.FullBleed);
+  render(template, block);
 
   // Update block with rendered template
   block.innerHTML = '';
   block.append(template);
-
-  const pictures = block.querySelectorAll('picture');
-  addPhotoCredit(pictures);
 
   // Inner block loading
   block.querySelectorAll('h2').forEach((h2) => {
@@ -152,11 +111,7 @@ ${!gdPlusArticle ? '' : premiumArticleBanner()}
   });
 
   block
-    .querySelectorAll('.social-share, .embed, .more-cards')
+    .querySelectorAll('.social-share, .embed')
     .forEach((innerBlock) => decorateBlock(innerBlock));
   loadBlocks(document.querySelector('main'));
-
-  if (gdPlusArticle) {
-    generateArticleBlocker(block, '.article-body');
-  }
 }
