+++ blocks/gallery/gallery.css
@@ -20,7 +20,7 @@
   background-color: #f2f2f2;
 }
 
-.gallery.block .container-article img:not(.gd-plus-icon) {
+.gallery.block .container-article img {
   width: 100%;
   height: auto;
   object-fit: contain;
