+++ blocks/header/header.html
@@ -40,7 +40,7 @@
             <a data-logged-in="false" class="profile-info-profile-photo" id="nav-photo"
                href="/my-account">
             </a>
-            <a class="login-dialog-open-button profile-info-button-login" href="#" data-type="gigya-login">Log In</class=>
+            <a class="profile-info-button-login" href="#" data-type="gigya-login">Log In</a>
         </div>
     </div>
 </div>
