+++ blocks/header/nav.html
@@ -39,8 +39,8 @@
         <div class="profile-info-overtitle">My Golf Digest Account</div>
         <ul class="profile-info">
           <li class="profile-info-button-wrapper">
-            <a class="login-dialog-open-button profile-info-button-login  o-button" href="#" data-type="idsp-login">Log In</a>
-            <a id="signup-button"class="profile-info-button-register o-button o-button-no-styles" href="#" data-type="idsp-register">Sign Up</a>
+            <a class="profile-info-button-login o-button" href="#" data-type="idsp-login">Log In</a>
+            <a class="profile-info-button-register o-button o-button-no-styles" href="#" data-type="idsp-register">Sign Up</a>
           </li>
         </ul>
     </div>
