+++ blocks/hero/hero.css
@@ -208,16 +208,6 @@
     background: #000;
 }
 
-.hero.block .button-container span {
-    font-size: 12px;
-}
-
-@media (min-width: 769px) {
-    .hero.block .button-container span {
-        font-size: 14px;
-    }
-}
-
 .hero.block .hero-cards-container {
     padding: 15px 18px;
     width: 100%;
@@ -285,10 +275,6 @@
     transition: transform 3s;
 }
 
-.hero.block.columns .hero-image-container img {
-    transition: transform 3s;
-}
-
 .hero.block .hero-cards-container a:hover img {
     transform: scale(1.2);
 }
@@ -325,132 +311,4 @@
     font-size: 16px;
     line-height: 130%;
     margin-top: 10px;
-}
-
-@media (max-width: 768px) {
-    .hero.block .hero-cards-container a .hero-card-title {
-        font-size: 14px;
-    }
-}
-
-/* hero columns variant */
-.hero.block.columns .hero-text-container {
-    grid-row: 2;
-    padding-bottom: 65px;
-    padding-top: 20px;
-    background-color: #ededed;
-    justify-content: flex-end;
-}
-
-.hero.block.columns .hero-container::after {
-    padding-bottom: 80%;
-}
-
-@media (min-width: 769px) {
-    .hero.block.columns .hero-container::after {
-        padding-bottom: 46%;
-    }
-}
-
-@media (min-width: 769px) {
-    .hero.block.columns .hero-text-container {
-        padding: 30px;
-        grid-row: 1;
-    }
-
-    .hero.block.columns.text-left .hero-text-container {
-        grid-column: 1/3;
-    }
-
-    .hero.block.columns.text-right .hero-text-container {
-        grid-column: 3/-1;
-    }
-}
-
-@media (min-width: 1281px) {
-    .hero.block.columns .hero-text-container {
-        padding: 100px;
-    }
-}
-
-@media (min-width: 769px) {
-    .hero.block.columns.with-cards .hero-text-container {
-        padding-bottom: 115px;
-    }
-}
-
-@media (min-width: 1281px) {
-    .hero.block.columns.with-cards .hero-text-container {
-        padding-bottom: 190px;
-    }
-}
-
-.hero.block.columns .hero-text-container a {
-    width: 100%;
-}
-
-.hero.block.columns :is(h1, h2) {
-    padding-bottom: 10px;
-}
-
-.hero.block.columns :is(h1, h2) + p {
-    padding-bottom: 20px;
-}
-
-.hero.block.columns :is(h1, h2),
-.hero.block.columns :is(h1, h2) + p,
-.hero.block.columns .hero-text-container .button-container {
-    color: #151517;
-    margin: 0;
-}
-
-@media (min-width: 1281px) {
-    .hero.block.columns :is(h1, h2) + p {
-        font-size: 16px;
-        padding-bottom: 25px;
-    }
-}
-
-@media (min-width: 769px) {
-    .hero.block.columns .hero-image-container {
-        grid-row: 1;
-    }
-
-    .hero.block.columns.text-left .hero-image-container {
-        grid-column: 3/-1;
-    }
-
-    .hero.block.columns.text-right .hero-image-container {
-        grid-column: 1/3;
-    }
-}
-
-.hero.block.columns .hero-image-container::after {
-    display: none;
-}
-
-.hero.block.columns:hover .hero-image-container img {
-    transform: scale(1.1);
-}
-
-/* Hero dark */
-.dark .hero.block .hero-text-container {
-    background-color: #151517;
-    color: #fff;
-}
-
-.dark .hero.block .hero-text-container a {
-    justify-content: center;
-}
-
-.dark .hero.block .hero-text-container :is(h1, h2, p, a, span) {
-    color: #fff;
-}
-
-.dark .hero.block .hero-container {
-    margin-top: 90px;
-}
-
-.dark .hero.block.columns .hero-container {
-    margin-top: 0;
 }
