+++ blocks/hero/hero.js
@@ -1,9 +1,5 @@
 import {
-  parseFragment,
-  render,
-  removeEmptyElements,
-  assignSlot,
-  getBlockId,
+  parseFragment, render, removeEmptyElements, assignSlot, getBlockId,
 } from '../../scripts/scripts.js';
 import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
 
@@ -16,13 +12,12 @@ const placeholderHtml = '<div class="hero-container" style="visibility: hidden; 
 
 export default async function decorate(block) {
   const id = getBlockId(block);
+  const query = window.store.getQuery(block);
 
-  const isFirstHero = document
-    .querySelector('.hero.block[data-block-name="hero"]')
-    .isEqualNode(block);
+  const isFirstHero = document.querySelector(".hero.block[data-block-name='hero']").isEqualNode(block);
 
   const heroDataIndex = heroItemsIndex;
-  heroItemsIndex += isFirstHero ? 5 : 1;
+  heroItemsIndex += (isFirstHero ? 5 : 1);
 
   // settings placeholder data
   if (!heroItems) {
@@ -34,87 +29,43 @@ export default async function decorate(block) {
     heroItems = event.detail.data;
 
     const heroData = heroItems[heroDataIndex];
-    // TODO Add support for multiple queries
     const cards = heroItems.slice(heroDataIndex + 1, heroDataIndex + 5);
 
     assignSlot(block, 'image', 'picture');
 
-    let cardsTemplate = '';
-
-    if (!block.classList.contains('no-cards')) {
-      cardsTemplate = cards
-        .map(
-          (card) => `
-              <a href='${card.path}'>
-                ${
-  createOptimizedPicture(card.image, card.imageAlt || card.title, isFirstHero, [{ width: '120' }])
-    .outerHTML
-}
+    const cardsTemplate = cards
+      .map((card) => `
+              <a href="${card.path}">
+                ${createOptimizedPicture(card.image, card.imageAlt || 'hero card image', isFirstHero, [{ width: '120' }]).outerHTML}
                 <div>
                   <span>${card.rubric}</span>
                   <p class="hero-card-title">${card.title}</p>
                 </div>
               </a>
-            `,
-        )
-        .join('');
-    }
-
-    if (!block.classList.contains('no-cards')) {
-      if (isFirstHero && cards.length) {
-        block.classList.add('with-cards');
-      } else {
-        block.classList.add('no-cards');
-      }
-    }
+            `)
+      .join('');
 
     // HTML template in JS to avoid extra waterfall for LCP blocks
     const HTML_TEMPLATE = `
-    <div class="hero-container">
-        <div class="hero-image-container">
-        ${
-  createOptimizedPicture(
-    heroData.image,
-    heroData.imageAlt || heroData.title,
-    isFirstHero,
-    [
-      { media: '(min-width: 686px)', width: '686' },
-      { media: '(min-width: 966px)', width: '966' },
-      { media: '(min-width: 1280px)', width: '1280' },
-      { width: '1850' },
-    ],
-  ).outerHTML
-}
+        <div class="hero-container">
+          <div class="hero-image-container">
+            ${createOptimizedPicture(heroData.image, heroData.imageAlt || 'hero main image', isFirstHero, [{ media: '(min-width: 686px)', width: '686' }, { media: '(min-width: 966px)', width: '966' }, { media: '(min-width: 1280px)', width: '1280' }, { width: '1850' }]).outerHTML}
+          </div>
+          <div class="hero-text-container">
+            <a href="${heroData.path}">
+              <div>
+                <${isFirstHero ? 'h1' : 'h2'}>${heroData.rubric}</${isFirstHero ? 'h1' : 'h2'}>
+                <p>${heroData.title}</p>
+              </div>
+              <div class="button-container">
+                <div class="icon-container">${arrowIcon}</div>
+                <span>Read story</span>
+              </div>
+            </a>
+          </div>
+          ${isFirstHero && cards.length ? `<div class="hero-cards-container">${cardsTemplate}</div>` : ''}
         </div>
-        <div class="hero-text-container">
-        <a href='${heroData.path}'>
-            <div>
-            <${isFirstHero ? 'h1' : 'h2'}>${
-  block.classList.contains('no-cards') ? heroData.title : heroData.rubric
-}</${isFirstHero ? 'h1' : 'h2'}>
-            ${
-  block.classList.contains('no-cards')
-    ? ''
-    : `<p>${heroData.title}</p>`
-}
-            </div>
-            <div class="button-container">
-            <div class="icon-container">${arrowIcon}</div>
-            <span>${
-  block.classList.contains('no-cards')
-    ? 'Watch series'
-    : 'Read story'
-}</span>
-            </div>
-        </a>
-        </div>
-        ${
-  cardsTemplate
-    ? `<div class="hero-cards-container">${cardsTemplate}</div>`
-    : ''
-}
-    </div>
-`;
+      `;
 
     // Parse the HTML template into a document fragment
     const template = parseFragment(HTML_TEMPLATE);
@@ -126,10 +77,7 @@ export default async function decorate(block) {
     removeEmptyElements(template, 'p');
 
     // Default text alignment
-    if (
-      !block.classList.contains('text-center')
-      && !block.classList.contains('text-right')
-    ) {
+    if (!block.classList.contains('text-center') && !block.classList.contains('text-right')) {
       block.classList.add('text-left');
     }
 
@@ -139,5 +87,5 @@ export default async function decorate(block) {
   });
 
   // Trigger query
-  window.store.query(block);
+  window.store.query({ id, query });
 }
