--- blocks/more-cards/more-cards.css
@@ -1,127 +1,0 @@
-.more-cards {
-    padding: 15px 10px;
-    margin: 30px 0;
-    display: flex;
-    justify-content: center;
-}
-
-.more-cards.multiple {
-    background-color: #f6f6f6;
-}
-
-.more-cards:not(.multiple) > div {
-    background-color: #f6f6f6;
-    padding: 10px;
-}
-
-@media (min-width: 1025px) {
-    .more-cards:not(.multiple) > div {
-        width: 510px;
-    }
-}
-
-@media (min-width: 1025px) {
-    .more-cards {
-        margin: 60px 0;
-    }
-}
-
-.more-cards.multiple > div > div {
-    display: flex;
-    justify-content: center;
-    width: 100%;
-    flex-wrap: wrap;
-    gap: 30px;
-}
-
-.more-cards h3 {
-    font-family: var(--font-gotham);
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 19px;
-    letter-spacing: .3em;
-    text-align: left;
-    text-transform: uppercase;
-    color: #989898;
-    margin-left: 8px;
-}
-
-.more-cards.multiple a {
-    display: flex;
-    flex-direction: column;
-    text-decoration: none;
-    width: 100%;
-    border: .5px solid #e5e5e5;
-    background-color: #fff;
-}
-
-@media (min-width: 1025px) {
-    .more-cards.multiple a {
-        width: 342px;
-    }
-}
-
-.more-cards picture {
-    margin: -0.5px;
-    width: calc(100% + 1px);
-    height: auto;
-    overflow: hidden;
-    position: relative;
-    display: block;
-}
-
-.more-cards:not(.multiple) picture {
-    padding: 10px;
-}
-
-.more-cards picture::before {
-    content: "";
-    display: block;
-    padding-top: 56.25%;
-    width: 100%;
-}
-
-.more-cards img {
-    position: absolute;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    transition: transform 3s;
-}
-
-.more-cards a:hover img {
-    transform: scale(1.1);
-}
-
-.more-cards span {
-    color: #adb4b7;
-    text-transform: uppercase;
-    font-weight: 400;
-    font-size: 10px;
-    line-height: 10px;
-    letter-spacing: .1em;
-    margin-bottom: 10px;
-    padding: 20px 20px 0;
-    display: block;
-}
-
-.more-cards:not(.multiple) span {
-    padding: 10px 10px 0;
-}
-
-.more-cards strong {
-    display: block;
-    color: #151517;
-    font-family: var(--font-gotham);
-    font-style: normal;
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 130%;
-    padding: 0 20px 20px;
-}
-
-.more-cards:not(.multiple) strong {
-    padding: 0 10px 10px;
-}
