--- blocks/series-cards/series-cards.css
@@ -1,105 +1,0 @@
-.series-cards.video-article.block {
-  margin-top: 74px;
-  width: 100%;
-  text-decoration: none;
-}
-
-.series-cards.video-article.block .series-cards-content {
-  padding: 0 15px;
-  max-width: 1440px;
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  gap: 60px 24px;
-  margin: 0 auto;
-}
-
-.series-cards.video-article.block .series-card-wrapper {
-  background-color: #1f2022;
-}
-
-.series-cards.video-article.block .card-link {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  text-decoration: none;
-}
-
-.series-cards.video-article.block .card-link:hover {
-  text-decoration: none;
-}
-
-.series-cards.video-article.block .series-card-wrapper .image-wrapper {
-  overflow: hidden;
-  width: 100%;
-  aspect-ratio: 16/9;
-}
-
-.series-cards.video-article.block .series-card-wrapper .image-wrapper picture {
-  height: 100%;
-  width: 100%;
-}
-
-.series-cards.video-article.block .series-card-wrapper .image-wrapper img {
-  height: 100%;
-  width: 100%;
-  object-fit: cover;
-  object-position: 50% 50%;
-  transition: transform 3s;
-  color: #fff;
-}
-
-.series-cards.video-article.block .series-card-wrapper .card-link:hover .image-wrapper img {
-  transform: scale(1.2);
-}
-
-.series-cards.video-article.block .series-card-wrapper .title-wrapper {
-  padding: 20px;
-}
-
-.series-cards.video-article.block .series-card-wrapper .title-span {
-  display: block;
-  color: #fff;
-  font-family: var(--font-gotham);
-  font-style: normal;
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 130%;
-  text-decoration: none;
-}
-
-@media (max-width: 1024px) {
-  .series-cards.video-article.block .series-cards-content {
-    grid-template-columns: 1fr 1fr;
-    grid-row-gap: 24px;
-    padding: 0;
-  }
-}
-
-@media (max-width: 768px) {
-  .series-cards.video-article.block .series-cards-content {
-    grid-template-columns: 1fr;
-    grid-row-gap: 1px;
-  }
-
-  .series-cards.video-article.block .card-link {
-    padding: 15px 18px 20px;
-    flex-direction: row;
-  }
-  
-  .series-cards.video-article.block .series-card-wrapper .image-wrapper {
-    overflow: hidden;
-    width: 132px;
-    height: 85px;
-    aspect-ratio: unset;
-  }
-
-  .series-cards.video-article.block .series-card-wrapper .title-wrapper {
-    padding-top: 0;
-    flex: 1;
-  }
-  
-  .series-cards.video-article.block .series-card-wrapper .title-span {
-    font-size: 14px;
-  }
-}
