+++ blocks/social-share/social-share.js
@@ -3,11 +3,6 @@ import { getMetadata } from '../../scripts/lib-franklin.js';
 /**
  * @param {HTMLDivElement} block
  */
-
-export const facebookSvg = '<svg viewBox="0 0 12 27" aria-label="Facebook"><path d="M8.26953125,8.72265625 L12.4042969,8.72265625 L11.9511719,13.3671875 L8.26953125,13.3671875 L8.26953125,26.6777344 L2.77539062,26.6777344 L2.77539062,13.3671875 L0,13.3671875 L0,8.72265625 L2.77539062,8.72265625 L2.77539062,6.00390625 C2.77539062,4.04035477 3.22851109,2.54883322 4.13476562,1.52929688 C5.07878076,0.509760527 6.60806234,0 8.72265625,0 L12.4042969,0 L12.4042969,4.58789062 L10.0820312,4.58789062 C9.66666459,4.58789062 9.32682424,4.62565066 9.0625,4.70117188 C8.83593637,4.77669309 8.66601619,4.90885322 8.55273438,5.09765625 C8.43945256,5.28645928 8.36393248,5.47525947 8.32617188,5.6640625 C8.28841127,5.85286553 8.26953125,6.1171858 8.26953125,6.45703125 L8.26953125,8.72265625 Z"></path></svg>';
-export const twitterSvg = '<svg viewBox="0 0 25 21" aria-label="Twitter"><path d="M25.4882812,2.43554688 C24.7708297,3.49284383 23.9023488,4.39908477 22.8828125,5.15429688 L22.8828125,5.83398438 C22.8828125,7.23112678 22.6751323,8.62824822 22.2597656,10.0253906 C21.844399,11.422533 21.2213583,12.7630144 20.390625,14.046875 C19.5598917,15.3307356 18.5592506,16.4635368 17.3886719,17.4453125 C16.2558537,18.4270882 14.8776123,19.2200491 13.2539062,19.8242188 C11.5924396,20.3906278 9.83659779,20.6738281 7.98632813,20.6738281 C5.0787615,20.6738281 2.41667875,19.8997473 0,18.3515625 C0.377606055,18.3893231 0.792966484,18.4082031 1.24609375,18.4082031 C3.6627725,18.4082031 5.81509473,17.6718824 7.703125,16.1992188 C6.57030684,16.1614581 5.5696658,15.8027378 4.70117188,15.1230469 C3.79491734,14.443356 3.1718767,13.5937551 2.83203125,12.5742188 C3.2096373,12.6119794 3.54947766,12.6308594 3.8515625,12.6308594 C4.30468977,12.6308594 4.75781023,12.5742193 5.2109375,12.4609375 C4.00259812,12.1966133 3.00195709,11.5924526 2.20898438,10.6484375 C1.41601166,9.70442236 1.01953125,8.59050121 1.01953125,7.30664062 L1.01953125,7.25 C1.73698275,7.66536666 2.52994357,7.89192689 3.3984375,7.9296875 C2.680986,7.43879963 2.11458541,6.81575898 1.69921875,6.06054688 C1.24609148,5.30533477 1.01953125,4.47461391 1.01953125,3.56835938 C1.01953125,2.62434424 1.2649715,1.75586334 1.75585938,0.962890625 C3.07748057,2.54883605 4.66340221,3.83267738 6.51367187,4.81445312 C8.40170215,5.75846826 10.4029842,6.28710881 12.5175781,6.40039062 C12.4420569,5.98502396 12.4042969,5.58854355 12.4042969,5.2109375 C12.4042969,3.77603449 12.9140574,2.54883322 13.9335938,1.52929688 C14.9531301,0.509760527 16.1803314,0 17.6152344,0 C19.1256586,0 20.4094999,0.547520566 21.4667969,1.64257813 C22.6373756,1.41601449 23.7324168,1.00065406 24.7519531,0.396484375 C24.3743471,1.60482375 23.6191463,2.54882473 22.4863281,3.22851562 C23.4681039,3.15299441 24.4687449,2.88867414 25.4882812,2.43554688 Z"></path></svg>';
-export const linkedInSvg = '<svg viewBox="0 0 16 16" aria-label="LinkedIn"><path d="M12.57 15.643v-5.518c0-1.387-.496-2.332-1.737-2.332-.947 0-1.513.638-1.759 1.254-.092.221-.114.527-.114.836v5.76H5.531s.045-9.345 0-10.312H8.96v1.46c.456-.703 1.272-1.703 3.091-1.703C14.309 5.087 16 6.562 16 9.729v5.914h-3.43zM1.918 3.921h-.022C.745 3.921 0 3.13 0 2.14 0 1.128.768.358 1.941.358c1.174 0 1.896.77 1.918 1.782 0 .99-.744 1.781-1.941 1.781zm1.715 11.722H.204V5.33h3.429v10.313z" fill="#FFF"></path></svg>';
-
 export default async function decorate(block) {
   const title = getMetadata('og:title');
   const url = getMetadata('og:url');
@@ -19,21 +14,27 @@ export default async function decorate(block) {
       <li>
         <a href="https://www.facebook.com/sharer.php?u=${url}" target="_blank" title="Share on Facebook">
           <span class="icon">
-            ${facebookSvg}
+            <svg viewBox="0 0 12 27" aria-label="Facebook">
+              <path d="M8.26953125,8.72265625 L12.4042969,8.72265625 L11.9511719,13.3671875 L8.26953125,13.3671875 L8.26953125,26.6777344 L2.77539062,26.6777344 L2.77539062,13.3671875 L0,13.3671875 L0,8.72265625 L2.77539062,8.72265625 L2.77539062,6.00390625 C2.77539062,4.04035477 3.22851109,2.54883322 4.13476562,1.52929688 C5.07878076,0.509760527 6.60806234,0 8.72265625,0 L12.4042969,0 L12.4042969,4.58789062 L10.0820312,4.58789062 C9.66666459,4.58789062 9.32682424,4.62565066 9.0625,4.70117188 C8.83593637,4.77669309 8.66601619,4.90885322 8.55273438,5.09765625 C8.43945256,5.28645928 8.36393248,5.47525947 8.32617188,5.6640625 C8.28841127,5.85286553 8.26953125,6.1171858 8.26953125,6.45703125 L8.26953125,8.72265625 Z"></path>
+            </svg>
           </span>
         </a>
       </li>
       <li>
         <a href="https://twitter.com/intent/tweet?text=${title}&url=${url}" target="_blank" title="Share on Twitter">
           <span class="icon">
-            ${twitterSvg}
+            <svg viewBox="0 0 25 21" aria-label="Twitter">
+              <path d="M25.4882812,2.43554688 C24.7708297,3.49284383 23.9023488,4.39908477 22.8828125,5.15429688 L22.8828125,5.83398438 C22.8828125,7.23112678 22.6751323,8.62824822 22.2597656,10.0253906 C21.844399,11.422533 21.2213583,12.7630144 20.390625,14.046875 C19.5598917,15.3307356 18.5592506,16.4635368 17.3886719,17.4453125 C16.2558537,18.4270882 14.8776123,19.2200491 13.2539062,19.8242188 C11.5924396,20.3906278 9.83659779,20.6738281 7.98632813,20.6738281 C5.0787615,20.6738281 2.41667875,19.8997473 0,18.3515625 C0.377606055,18.3893231 0.792966484,18.4082031 1.24609375,18.4082031 C3.6627725,18.4082031 5.81509473,17.6718824 7.703125,16.1992188 C6.57030684,16.1614581 5.5696658,15.8027378 4.70117188,15.1230469 C3.79491734,14.443356 3.1718767,13.5937551 2.83203125,12.5742188 C3.2096373,12.6119794 3.54947766,12.6308594 3.8515625,12.6308594 C4.30468977,12.6308594 4.75781023,12.5742193 5.2109375,12.4609375 C4.00259812,12.1966133 3.00195709,11.5924526 2.20898438,10.6484375 C1.41601166,9.70442236 1.01953125,8.59050121 1.01953125,7.30664062 L1.01953125,7.25 C1.73698275,7.66536666 2.52994357,7.89192689 3.3984375,7.9296875 C2.680986,7.43879963 2.11458541,6.81575898 1.69921875,6.06054688 C1.24609148,5.30533477 1.01953125,4.47461391 1.01953125,3.56835938 C1.01953125,2.62434424 1.2649715,1.75586334 1.75585938,0.962890625 C3.07748057,2.54883605 4.66340221,3.83267738 6.51367187,4.81445312 C8.40170215,5.75846826 10.4029842,6.28710881 12.5175781,6.40039062 C12.4420569,5.98502396 12.4042969,5.58854355 12.4042969,5.2109375 C12.4042969,3.77603449 12.9140574,2.54883322 13.9335938,1.52929688 C14.9531301,0.509760527 16.1803314,0 17.6152344,0 C19.1256586,0 20.4094999,0.547520566 21.4667969,1.64257813 C22.6373756,1.41601449 23.7324168,1.00065406 24.7519531,0.396484375 C24.3743471,1.60482375 23.6191463,2.54882473 22.4863281,3.22851562 C23.4681039,3.15299441 24.4687449,2.88867414 25.4882812,2.43554688 Z"></path>
+            </svg>
           </span>
         </a>
       </li>
       <li>
         <a href="https://www.linkedin.com/sharing/share-offsite/?url=${url}" target="_blank" title="Share on LinkedIn">
           <span class="icon">
-            ${linkedInSvg}
+            <svg viewBox="0 0 16 16" aria-label="LinkedIn">
+                <path d="M12.57 15.643v-5.518c0-1.387-.496-2.332-1.737-2.332-.947 0-1.513.638-1.759 1.254-.092.221-.114.527-.114.836v5.76H5.531s.045-9.345 0-10.312H8.96v1.46c.456-.703 1.272-1.703 3.091-1.703C14.309 5.087 16 6.562 16 9.729v5.914h-3.43zM1.918 3.921h-.022C.745 3.921 0 3.13 0 2.14 0 1.128.768.358 1.941.358c1.174 0 1.896.77 1.918 1.782 0 .99-.744 1.781-1.941 1.781zm1.715 11.722H.204V5.33h3.429v10.313z" fill="#FFF"></path>
+            </svg>
           </span>
         </a>
       </li>
