--- blocks/text-box/text-box.css
@@ -1,76 +1,0 @@
-.text-box.block {
-  max-width: 1024px;
-  margin: 0 auto;
-}
-
-.text-box.block .text-box-content {
-  padding: 40px 80px;
-  background-color: rgb(223 223 223);
-  color: #000;
-  margin: 0 15px 14px;
-}
-
-.text-box.block.dark .text-box-content {
-  background-color: #1f2022;
-  color: #fff;
-}
-
-.text-box.block .text-box-content * {
-  margin: 0;
-}
-
-.text-box.block .text-box-content h1,
-.text-box.block .text-box-content h2,
-.text-box.block .text-box-content h3 {
-  font-weight: 600;
-  display: flex;
-  align-items: center;
-  font-size: 28px;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-  font-family: var(--font-tungsten);
-  margin-bottom: 25px;
-}
-
-.text-box.block .text-box-content h1::before,
-.text-box.block .text-box-content h2::before,
-.text-box.block .text-box-content h3::before {
-  content: "";
-  display: block;
-  width: 5px;
-  height: 20px;
-  margin-right: 10px;
-  background-color: #ed3e49;
-}
-
-.text-box.block .text-box-content p {
-  text-align: left;
-  font-size: 16px;
-  line-height: 25.6px;
-  font-family: var(--font-gotham);
-}
-
-@media (max-width: 1280px) {
-  .text-box.block .text-box-content h1,
-  .text-box.block .text-box-content h2,
-  .text-box.block .text-box-content h3 {
-    font-size: 24px;
-    padding: 0 15px;
-  }
-  
-  .text-box.block .text-box-content p {
-    margin: 0 24px;
-  }
-}
-
-@media (max-width: 1024px) {
-  .text-box.block .text-box-content {
-    padding: 20px 15px;
-  }
-}
-
-@media (max-width: 768px) {
-  .text-box.block .text-box-content {
-    margin: 0 15px 14px;
-  }
-}
