+++ helix-query.yaml
@@ -27,13 +27,6 @@ indices:
         select: head > meta[name="publication-date"]
         value: |
           attribute(el, 'content')
-      gdPlus:
-        select: head > meta[name="gd-plus"]
-        value: |
-          attribute(el, 'content')
-      lastModified:
-        select: none
-        value: parseTimestamp(headers["last-modified"], "ddd, DD MMM YYYY hh:mm:ss GMT")
   article:
     include:
      - '/content-v2/**/article/**'
@@ -71,17 +64,6 @@ indices:
         select: head > meta[name="author-url"]
         value: |
           attribute(el, 'content')
-      gdPlus:
-        select: head > meta[name="gd-plus"]
-        value: |
-          attribute(el, 'content')
-      videoUrl:
-        select: 'a[href*="youtube.com/watch"], a[href*="youtu.be/"], a[href*="brightcove"]'
-        value: |
-          attribute(el, 'href')
-      lastModified:
-        select: none
-        value: parseTimestamp(headers["last-modified"], "ddd, DD MMM YYYY hh:mm:ss GMT")
   products:
     include:
      - '/content-v2/**/products/**'
@@ -103,6 +85,7 @@ indices:
         select: head > meta[name="category"]
         value: |
           attribute(el, 'content')
-      lastModified:
-        select: none
-        value: parseTimestamp(headers["last-modified"], "ddd, DD MMM YYYY hh:mm:ss GMT")
+      date:
+        select: head > meta[name="publication-date"]
+        value: |
+          attribute(el, 'content')
