+++ scripts/scripts.js
@@ -117,7 +117,7 @@ export function render(template, fragment, type) {
       .replaceAll('</a> **', '</a></strong>')
       .replaceAll('from left to right:', 'from left to right:<br>')
       .replaceAll('</a><strong>', '</a><br><strong>');
-  } else if (type === ARTICLE_TEMPLATES.LongForm || type === ARTICLE_TEMPLATES.FullBleed) {
+  } else if (type === ARTICLE_TEMPLATES.LongForm) {
     fragment.innerHTML = fragment.innerHTML
       .replaceAll('****', '')
       .replaceAll('</em> **', '</em></strong>&nbsp;')
@@ -159,18 +159,18 @@ export function normalizeAuthorURL(author) {
  * @param el
  */
 export function addPortraitClass(el) {
-  if (el.tagName === 'PICTURE') {
-    const img = el.querySelector('img');
-    if (img && img.height > img.width) {
-      el.classList.add('portrait');
-    }
-  } else {
+  if (el.length) {
     el.forEach((picture) => {
       const img = picture.querySelector('img');
       if (img && img.height > img.width) {
         picture.classList.add('portrait');
       }
     });
+  } else {
+    const img = el.querySelector('img');
+    if (img && img.height > img.width) {
+      el.classList.add('portrait');
+    }
   }
 }
 
@@ -206,50 +206,6 @@ function buildTemplate(main) {
     if (template) {
       main.querySelectorAll('.section-metadata').forEach((metadataEl) => {
         metadataEl.className = 'template-section-metadata';
-      });
-
-      // TODO remove once importer fixes more cards
-      const checkForMoreCards = (el, elems) => {
-        if (el.tagName === 'P' && el.querySelector('picture') && el.querySelector('a') && el?.nextElementSibling?.tagName === 'P' && el.nextElementSibling.children[0]?.tagName === 'A' && el.nextElementSibling?.nextElementSibling.tagName === 'P' && el.nextElementSibling.nextElementSibling.children[0]?.tagName === 'A') {
-          const rubric = document.createElement('span');
-          rubric.textContent = el.nextElementSibling.textContent.trim();
-
-          const desc = document.createElement('strong');
-          desc.textContent = el.nextElementSibling.nextElementSibling.textContent.trim();
-
-          const link = document.createElement('a');
-          link.setAttribute('href', new URL(el.querySelector('a').getAttribute('href')).pathname);
-
-          link.append(el.querySelector('picture'));
-          link.append(rubric);
-          link.append(desc);
-
-          el.nextElementSibling.nextElementSibling.classList.add('remove');
-          el.nextElementSibling.classList.add('remove');
-          el.classList.add('remove');
-
-          elems.push(link);
-
-          if (el.nextElementSibling.nextElementSibling.nextElementSibling) {
-            checkForMoreCards(el.nextElementSibling.nextElementSibling.nextElementSibling, elems);
-          }
-        }
-      };
-
-      main.querySelectorAll('h2').forEach((h2) => {
-        if (h2.nextElementSibling) {
-          const elems = [];
-          checkForMoreCards(h2.nextElementSibling, elems);
-          if (elems.length) {
-            main.querySelectorAll('.remove').forEach((el) => el.remove());
-            const h3 = document.createElement('h3');
-            h3.textContent = h2.textContent;
-            h3.id = h2.id;
-            elems.unshift(h3);
-            const moreCards = buildBlock('more-cards', { elems });
-            h2.replaceWith(moreCards);
-          }
-        }
       });
 
       const section = document.createElement('div');
@@ -430,7 +386,7 @@ export const convertExcelDate = (excelDate) => {
 };
 
 export function getBlockId(block) {
-  block.id = block.id || window.crypto.randomUUID();
+  block.id = window.crypto.randomUUID();
   return block.id;
 }
 
@@ -439,43 +395,43 @@ export function getBlockId(block) {
  */
 window.store = new (class {
   constructor() {
-    // Query stack
-    this._queryStack = {};
+    // Fetch stack
+    this._stack = {};
 
     // Indexes
     this._spreadsheets = {
-      article: 'article-query-index',
-      product: 'product-query-index',
-      gallery: 'gallery-query-index',
-      'custom-data': 'custom-data',
+      ARTICLES: 'article-query-index',
+      PRODUCTS: 'product-query-index',
+      GALLERY: 'gallery-query-index',
     };
 
-    // Pre-defined queries
+    // Author query to query map
     this._queryMap = {
+      news: {
+        spreadsheet: this._spreadsheets.ARTICLES,
+        sheet: 'golf-news-tours-default',
+      },
+      features: {
+        spreadsheet: this._spreadsheets.ARTICLES,
+        sheet: 'golf-news-tours-features',
+      },
+      // TODO query gd+ articles
+      gd: {
+        spreadsheet: this._spreadsheets.ARTICLES,
+        sheet: 'golf-news-tours-features',
+      },
+      courses: {
+        spreadsheet: this._spreadsheets.ARTICLES,
+        sheet: 'courses',
+      },
+      loop: {
+        spreadsheet: this._spreadsheets.ARTICLES,
+        sheet: 'loop',
+      },
       wedges: {
         mock: '/mock-data/wedges.json',
-        limit: 0,
       },
     };
-
-    // Max items per block
-    /*
-    * Add entry here for new blocks that require data to be fetched!
-    */
-    this._blockQueryLimit = {
-      hero: 5,
-      cards: 10,
-      carousel: 20,
-      loop: 30,
-      'series-cards': 100,
-      'tiger-cards': 35,
-      'tiger-vault-hero': 1,
-    };
-
-    this.blockNames = Object.keys(this._blockQueryLimit);
-    this.spreadsheets = Object.keys(this._spreadsheets);
-
-    this.initQueries();
 
     try {
       this._cache = JSON.parse(window.sessionStorage['golf-store']);
@@ -488,36 +444,6 @@ window.store = new (class {
       // session storage not supported
       this._cache = {};
     }
-  } // Find all dynamic queries on the page and map them out
-
-  // Dynamic queries end with their spreadsheet name e.g. loop-article or latest-gallery
-  initQueries() {
-    const dynamicQuerySelectors = this.spreadsheets.map((spreadsheet) => `main .block[class*="${spreadsheet}"]`);
-    document.querySelectorAll(dynamicQuerySelectors.join(',')).forEach((block) => {
-      for (const spreadsheet of this.spreadsheets) {
-        const queryClassName = [...block.classList]
-          .find((className) => className.endsWith(spreadsheet));
-
-        if (queryClassName) {
-          const query = queryClassName.replace(`-${spreadsheet}`, '');
-          if (!this._queryMap[query]) {
-            this._queryMap[query] = {
-              spreadsheet,
-              limit: 0,
-            };
-          }
-        }
-      }
-    });
-
-    // Construct limit for each query based on the blocks on the page
-    const queryNames = Object.keys(this._queryMap);
-    this.blockNames.forEach((blockName) => {
-      queryNames.forEach((queryName) => {
-        const { spreadsheet } = this._queryMap[queryName];
-        this._queryMap[queryName].limit += document.querySelectorAll(`main .${blockName}.${queryName}-${spreadsheet}.block`).length * this._blockQueryLimit[blockName];
-      });
-    });
   }
 
   /**
@@ -526,88 +452,68 @@ window.store = new (class {
    * @return {string}
    */
   getQuery(block) {
-    const queries = Object.keys(this._queryMap);
+    let query = '';
 
-    for (const query of queries) {
-      const found = [...block.classList].find((className) => className.startsWith(query));
-      if (found) {
-        block.dataset.query = query;
+    const queries = Object.keys(this._queryMap);
+    for (const className of block.classList) {
+      if (queries.includes(className)) {
+        query = className;
         break;
       }
     }
 
-    return block.dataset.query;
+    return query;
   }
 
   /**
    * Triggers an index fetch
-   * @param {HTMLElement} block
+   * @param {object}
    */
-  query(block) {
-    const id = getBlockId(block);
-    let query = this.getQuery(block);
-
+  query({
+    id, query, offset = 0, limit = 10,
+  }) {
     if (!query) {
-      // Attempt to find more blocks generated late.
-      // This usually happens when using a non-default page-
-      // type as you have to trigger block decoration manually.
-      this.initQueries();
-      query = this.getQuery(block);
-      if (!query) {
-        console.warn(`Query missing for "${block.dataset.blockName}" with id "${id}"!
-Make sure the block definition includes the spreadsheet & sheet name like this: \x1b[37m"(example, class, SHEET_NAME SPREADSHEET_FILENAME)"\x1b[0m!
-If you created a new spreadsheet you might also need to add it to \x1b[37m"this._spreadsheets"\x1b[0m.
-`);
-        return;
-      }
+      console.warn(`Query missing for "${id}"`);
+      return;
     }
 
+    // Check if query is available
     const queryDetails = this._queryMap[query];
-    if (!queryDetails.mock && queryDetails.limit < 1) {
-      console.warn(`No query limit was found for ${block.dataset.blockName} block! \x1b[1m\x1b[31mTherefore no data will be returned!\x1b[0m
-
-Make sure to set a limit for \x1b[31m"${block.dataset.blockName}"\x1b[0m in \x1b[37m${JSON.stringify(this._blockQueryLimit)}\x1b[0m
-      `);
+    if (!queryDetails) {
+      console.warn(`Query "${query}" not implemented`);
+      return;
     }
-    let url;
 
+    let url;
     // Use mock data if defined
     if (queryDetails.mock) {
       url = queryDetails.mock;
     } else {
       // Build query sheet url
-      url = `/${this._spreadsheets[queryDetails.spreadsheet]}.json?sheet=${query}`;
+      url = `/${queryDetails.spreadsheet}.json?offset=${offset}&limit=${limit}&sheet=${queryDetails.sheet}`;
     }
 
-    // Use cached resource & that it has data
-    if (this._cache[url] && this._cache[url].limit) {
-      // Cache is already populated
+    // Use cached resource
+    if (this._cache[url]) {
+      // Only trigger if there is data
       if (this._cache[url].data.length) {
-        // Only trigger if there is enough data
-        if (queryDetails.limit <= this._cache[url].data.length) {
-          // "Return" data for given id
-          document.dispatchEvent(new CustomEvent(`query:${id}`, { detail: this._cache[url] }));
-          return;
-        }
+        // "Return" data for given id
+        document.dispatchEvent(new CustomEvent(`query:${id}`, { detail: this._cache[url] }));
       } else {
         // Stack query
-        this._queryStack = {
-          ...this._queryStack,
-          [id]: url,
+        this._stack = {
+          ...this._stack,
+          id: url,
         };
-
-        return;
       }
+
+      return;
     }
 
-    // Start setting cache to avoid multiple requests or invalid cache if not enough items
+    // Start setting cache to avoid multiple requests
     this._cache[url] = { data: [] };
 
-    // TODO store the delta between the number of cached items and the number of items requested
-    // and only request that with ?offset=
-
-    // Fetch new data, cache it then trigger
-    fetch(queryDetails.mock ? url : `${url}&limit=${queryDetails.limit}`)
+    fetch(url)
       .then((req) => {
         if (req.ok) {
           return req.json();
@@ -629,87 +535,15 @@ Make sure to set a limit for \x1b[31m"${block.dataset.blockName}"\x1b[0m in \x1b
         document.dispatchEvent(new CustomEvent(`query:${id}`, { detail: this._cache[url] }));
 
         // Unstack and "return" data
-        for (const stackId in this._queryStack) {
-          if (url === this._queryStack[stackId]) {
+        Object.entries(this._stack).forEach((stackId, stackURL) => {
+          if (url === stackURL) {
             document.dispatchEvent(new CustomEvent(`query:${stackId}`, { detail: this._cache[url] }));
-            // Pop stack id
-            delete this._queryStack[stackId];
+            delete this._stack[stackId];
           }
-        }
+        });
       })
       .catch((error) => {
         console.warn(error);
       });
   }
 })();
-
-/**
- * Generates HTML for the premium article banner.
- * @param {Number} Number of leftover articles to compare to.
- */
-export const premiumArticleBanner = (customLeftoverArticles = null) => {
-  let leftoverArticles = customLeftoverArticles;
-  if (typeof customLeftoverArticles !== 'number') {
-    leftoverArticles = Math.min(Number(window.name), 3);
-  }
-  window.name = Math.max(leftoverArticles - 1, 0);
-
-  let text;
-  if (leftoverArticles > 1) {
-    text = `You have <strong>${leftoverArticles}</strong> free premium articles remaining.`;
-  }
-  if (leftoverArticles === 1) {
-    text = 'This is your last free premium article for the month.';
-  }
-  if (leftoverArticles < 1) {
-    text = 'You are out of free premium articles.';
-  }
-
-  return `
-    <div class="premium-article-banner ${leftoverArticles < 1 ? 'out-of-free-articles' : ''}">
-      <div class="premium-banner-text-wrapper">
-        <span class="premium-message">${text}</span>
-        <a class="premium-link" src="#" target="_blank">
-          Subscribe to <strong>Golf Digest<span class="red-plus">+</span></strong>
-        </a>
-      </div>
-    </div>
-  `;
-};
-/**
- * Generates HTML for the premium article blocker.
- * @param {block} Block where the selector exists.
- * @param {Selector} Selector for article body that should be covered.
- */
-export const generateArticleBlocker = (block, selector) => {
-  if (Number(window.sessionStorage.freeArticles) > 0) {
-    return;
-  }
-  const articleBody = block.querySelector(selector);
-
-  articleBody.style.height = '1000px';
-  articleBody.style.position = 'relative';
-  articleBody.style.overflow = 'hidden';
-
-  const articleBlocker = document.createElement('div');
-  articleBlocker.className = 'article-blocker-wrapper';
-  articleBlocker.innerHTML = `
-    <div class="article-blocker-content">
-      <img class="article-blocker-image gd-plus-logo" src="/icons/gd-plus-logo.svg" alt="Golf Digest Plus Logo" />
-      <div class="article-blocker-lead">Subscribe to continue Reading</div>
-      <div class="article-blocker-sublead">
-        <span class="highlight">Golf Digest<span class="red-plus">+</span></span>
-        includes unlimited digital articles, exclusive course reviews, magazine access and more!
-      </div>
-      <a class="cta" href="https://www.golfdigest.com/subscribe-golf-digest-plus" target="_blank">
-        Learn More
-      </a>
-
-      <span class="login-wrapper">
-      Already have an account? <button class="login-button" onclick="console.warn('login popup not implemented yet');">Log in</button>
-      </span>
-    </div>
-  `;
-
-  articleBody.appendChild(articleBlocker);
-};
